Take min dimensions in DP for SplitPairRule/SplitPlaceholderRule
Before, we take minWidth, minHeight, minSmallestWidth in pixels, which
doesn't get updated when the display density is changed. Now, we take
them in DP, so that we can convert it to pixels later with up-to-date
density.

Also, refactor to take minWidth, minHeight, minSmallestWidth as
optional for SplitRule Builder so that:
1. Allow the app to use the library default value if not set.
2. Allow us to introduce more bounds related options in the future.

Relnote: "
1. Refactor SplitRule#minWidth/minHeight/minSmallestWidth to
   SplitRule#minWidthDp/minHeightDp/minSmallestWidthDp to take DP
   instead of pixels.
2. For SplitPairRule/SplitPlaceholderRule Builder, introduce
   Builder#setMinWidthDp, Builder#setMinHeightDp and
   Builder#setMinSmallestWidthDp to take min dimensions as optional.
"
Bug: 229656253
Test: pass existing

Change-Id: I95f17ce89844d18708e6fb03924d38a00af115ac