Migrate Relation and RelationCollectorFunctionWriter to XPoet
Migrate the relationship code writers of @Relation fields to XPoet to generate Kotlin code. This CL refactors the writer to accommodate for Kotlin's nullability checks in types along with moving some of the generated code to runtime helpers.

First the nullability of the parent and child relation columns have to be accounted when generating code that reads the 'key' column. The solution implemented is to use column readers when reading keys as these are null-aware as opposed to the hardcoded cursor get functions used. The readers are always available since the key type is based on the matching fields affinity and not the actual field type and thus they are all built-in supported types (String, Long, etc). One special type is ByteBuffer which is supported via a built-in converter for BLOB affinity. To support always finding the converter Context.fork() was updated to receive forced built-in converter states.

Second, the generated code to recursively fetch relationships when exceeding the max binding args amount was moved to runtime helpers functions in RelationUtil. This makes them easier to maintain and reduces the amount of code generation logic in the writer. However, the fetch functions are a bit redundant with each other as they need to support various map types, specifically HashMap, LongSparseArray and ArrayMap. This also means room-runtime has a new compileOnly dependency on androidx.collection to support having these runtime helper functions. The function to be use is still determine by room-compiler and the availability of androidx.collection types in the project's classpath.

XProcessing change:
* Fixed requireType(XTypeName) in KSP to not simply use KTypeName as String since that might contain the `?` suffix when nullable and will fail lookup. Instead find the type using the canonical name and then make nullable or non null based on XTypeName.nullability.

New XPoet API:
* XCodeBlock.indent() and XCodeBlock.unindent() - For adding and removing code indent, useful for lambdas.
* XTypeName.rawTypeName - For getting the raw type name of a parametrized type name. Even though there is no XParametrizedTypeName,the function is available to all type names returning itself if not parametrized.

Bug: 127483380
Test: DaoRelationshipKotlinCodeGenTest + existing
Change-Id: I2baf26703b1cc57e14af23e7ac452028a389d9ed