Remove FullLifecycleObserver
Now that DefaultLifecycleObserver no longer
extends on FullLifecycleObserver, it can be
deleted. We want to keep all the functionality,
so we should rename the adapter and test suite
to "default".

RelNote: "`DefaultLifecycleObserver` now
extends the public `LifecycleObserver` directly
and `FullLifecycleObserver` has been deleted,
transferring all of its functionality to
`DefaultLifecycleObserver`."
Bug: 257137635
Test: all tests pass

Change-Id: Ia8457b389ea7274c6e18b457bb9505b1c31b62e8