Change announcement on 'isPassword' toggle
This changes the announcement when we toggle the password visibility icon in the material TextField.
This change follows the View-based implementation: it sends the TYPE_VIEW_TEXT_SELECTION_CHANGED. In View there's another TYPE_VIEW_TEXT_SELECTION_CHANGED event with fromIndex=0 and toIndex=0 which most likely come from setting the span. We don't send this event in Compose (only when the real selection changes). Therefore to mirror the behavior we need to send it. As a result two TYPE_VIEW_TEXT_SELECTION_CHANGED events are sent - one with the cursor position at 0 and second with the correct position.

Bug: 247891690
Test: manually in demo with Talkback on
Test: new tests in AndroidComposeViewAccessibilityDelegateCompatTest

Change-Id: If75a1d48df37a0475367aacfca925d39084d38f1