Fix throwable parcelable definition in older AIDL versions.
We don't have a way to define recursive parcelables in a way that's
compatible with AIDL versions 30 and 33. Before version 33, recursive
parcelables could just use the @nullable annotation, but after version
33 we are required to add the heap=true option to the annotation.

In this change we are instead modeling the throwable cause as a list
that can either be empty or contain a single cause.

Bug: 257417646
Test: ./gradlew privacysandbox:tools:tools-core:test
Change-Id: I757d4c0e632680ece95e9a70317ec6a25abf6e7f