Fix issue with XAnnotationValue#getValue() types.
This CL ensures that the `XAnnotationValue#getValue()` type matches the
type defined at the annotation's declaration site rather than the type
input at the usage site. This is already the case for Javac, but KSP
returns the exact value at the usage site, which may be a different
type than the declaration site.

For example, if the annotation value's type at the declaration site is
an `int`, then Javac's `AnnotationValue#getValue()` will return an `int`,
even if you pass a `byte` or `short` at the usage site (which is allowed
in Java source but not Kotlin source). However, KSP returns the type
exactly as declared in the usage site -- e.g. as a `byte` if it was input
as a byte at the usage site. This CL fixes this inconsistency by
detecting the type difference and manually converting the value to the
correct type.

Test: XAnnotationValueTest
Change-Id: I4926b54352f4fa2dd0de573d5acfc4887bac0789