Nested Fling and FlingBehavior issue fix
In certain cases it is necessary to let a parent scrollable to cancel the fling animation that is driven by a child scrollable through the nested scroll cycle. We'll chance the fling animation method to react to cancellation exceptions, in any case a parent wants to cancel the fling animation they can throw a cancellation exception and the fling will be cancelled.

Relnote: N/A

Test: Added tests
Fixes: 258796877
Fixes: 281676206
Fixes: 266829796

Change-Id: Id3d7988b32a161e607a34bbd47661c3a535037ef