Add switcher to the package page for *Kt files
If a java page does not have a corresponding kotlin page, this adds a
switcher to the package summary page as the extension functions are
displayed there in kotlin. This doesn't do the same for kotlin pages
because if there isn't a java page, it's because the package does not
exist for java.

Bug: 251807365
Test: cl/488357957, with staged changes at go/dac-stage/reference/androidx/packages?db=androidx-compose
Change-Id: Ib9426a2e09b7c7ec01e6cb43a4a50af92527f4e1