Add items pinning functionality to the lazy lists
Lazy lists now provide an instance of PinnableContainer via LocalPinnableContainer for each item. It allows to pin the current item from inside the content of the item. Pinning means we will continue composing, measuring and placing this item even when it is scrolled away from the view. Note that the placed position is not guaranteed to be the real one as we do not compose the items in between the pinned one and the first/last visible one.
This interface was added to the ui module as conceptually PinnableContainer is decoupled from lazy layouts from the foundation module. Other layouts using SubcomposeLayout could have some similar functionality and may want to support pinning. And it will allow for modifiers from ui to use this interface to pin their content.

Test: LazyListPinnableContentTest
Fixes: 259274257
Fixes: 195049010
Relnote: Introduced PinnableContainer api propagated by lazy lists via a composition local which allows to pin current item. This means such item will not be disposed when it is scrolled away from the view. For example, Modifier.focusable() will pin the current focused item via this mechanism.
Change-Id: Ib8881191a529c9d9dc5e886570650b1987763207