Accessibility fixes for picker
Add `userScrollEnabled` parameter in `Picker` which enables/disables the
scrolling. This parameter allows to make the `Picker` more accessible by
disabling the scrolling when the `Picker` is in `readOnly` mode. The
paramter `readOnly` and `userScrollEnabled` need to be different to
control the behaviour of `Picker` based on talkback and non-talkback
mode.

`userScrollEnabled` parameter comes from `LazyList` which has a similar
effect on the list.

Note: The `updateApi` build was failing because of a known bug in
Metalava (b/219792969), hence this change was updated with
`ignoreApiChange && updateApi`.

Bug: 255553263
Relnote: "Accessbility improvements for Picker for screens with
multi-picker. Allows to scroll the picker from anywhere on the screen
without triggering scroll on other pickers"
Test: ./gradlew :wear:compose:compose-material:ignoreApiChange &&
./gradlew :wear:compose:compose-material:updateApi
Test: Manual Testing

Change-Id: I3c3aa55ce84d0d577f3f7ab53a23be915abbc5a1