Refactor some animation tests to be more imperative and deterministic.
AnimatedContentTest.AnimatedContentSlideInAndOutOfContainerTest:
The assertions were written inside a LaunchedEffect for each branch of
the transition. The test actually runs two transitions in a row, with
different assertions for each run, but those were all embedded in the
same LaunchedEffect and triggered by comparing and setting some state in
a recomposition. This made the test hard to follow and update. I pulled
the state writes that trigger the animations, as well as the assertions,
out and used explicit advanceByFrame calls.

AnimatedVisibilityTest.testTransitionExtensionAnimatedVisibility:
The test was using tween animations with the default easing, which is
harder to write assertions around because due to the curve the animation
can reach final values before the animation duration is complete. I
changed the animation to use LinearEasing instead, and updated the test
assertions accordingly (e.g. the content will be completely removed from
the composition when hidden instead of being measured with zero size).

Other changes are using Truth for assertions and auto-formatting from
Android Studio.

Test: AnimatedContentTest, AnimatedVisibilityTest
Change-Id: Iee546c24ab1ead7805968e3cf5a92ca6df6fcf10