Resolving NPE issue in SupportSQLiteQueryBuilder.
This issue was caused by the incorrect logic of using `columns?.size != 0` on a nullable `column` to decide whether the contents of `column` can be appended to a statement. The logic was flawed in that the expression would return true even if `columns` itself was null, attempting to append non-existent content. The correct logic to use is `!columns.isNullOrEmpty()` which covers both cases. Same logic has been applied to `StringBuilder.appendClause` to avoid any potential issues with a null `clause`.

Bug: 260625714
Test: SupportSQLiteQueryBuilderTest.kt
Change-Id: Ica8f50febd3648b7539fa94df5f0a0197c94e63d