Use BringIntoViewResponder to bring element into view.
By default, scrollable-containers ensure that the focused element
is brought into view. We need to bring the entire FeaturedCarousel
or ImmersiveList into view even when a sub-composable is in focus.

The earlier approach called BringIntoViewRequester.bringIntoView.
This change uses a custom BringIntoViewResponder to intercept the
request to bringIntoView and return the entire FeaturedCarousel
or ImmersiveList rectangle rather than only the rectangle of the
focused sub-composable.

Test: Existing tests cover this change.
Relnote: NA
Change-Id: I232436a2c7ee3a72fdac762463c5167470964400