Make more of kotlinx.collections.immutable stable
This addresses two issues around stability inferencing for types in the
kotlinx.collections.immutable library. We have a few exceptions to treat
ImmutableList, ImmutableSet, and ImmutableMap as stable types, but we
don't have anything for ImmutableCollection or their Persistent
siblings. This CL adds those exceptions.

We also haven't marked any of their corresponding `*Of()` builders as
stable, which this CL also adds exceptions for.

Bug: 238057117
Bug: 254435410
Test: ClassStabilityTransformTests
Change-Id: Ie19290bfb53dd667e866681f11f266ae7bdb6201