[GH] Fix up core dependencies on annotation
We are inconsistently including parts of annotation as project or snapshot dependencies. There also seems to be a bug with configuration caching and task dependency, which is causing annotations to not get built. This change aligns all annotation deps to be projectOrArtifact, except for annotation-sampled, which is not published.

Test: cd core && ./gradlew bOS
Change-Id: Ie606e42d48a0f88fb8f32fa38357fea28045b74d

This is an imported pull request from https://github.com/androidx/androidx/pull/470.

Resolves #470
Github-Pr-Head-Sha: c7d086ca776757537ed2856e47e49af7cb52c65e
GitOrigin-RevId: 4f5bb96c0f9d9ba094008149075da799be0855be
Change-Id: Id2c8a6d19c4b2c20ffa76abdb7b7a7ec0ad7eed6