Improving < 24 API Changes and ErrMessages
Making changes on how we name shuttled keys through the result receiver,
how we utilize the result receiver in the hidden activity, cleaning code
that is not required, and improving error messaging across the codebase.
Note that PKC specific error messages will be added on in future
changes, but these cover error messages across the board generally.

Bug: 261092609
Test: Conversion Tests will be patched in, Builds, and Demos verify
Change-Id: I6ead3224e1eb3df2e8e0728224958539512b3faf