Adjust PagerFlowSnapshot initial lastAccessedIndex calculation
Upon a Refresh load, Paging would attempt to trigger prefetches on either direction if the [center index of initial load] - [presentedItemsBeyondAnchor] < prefetchDistance of that direction. This "anchor" essentially is the center index of the initial loaded items with placeholders accounted for.

In PagerFlowSnapshot, the initial lastAccessedIndex was based on index of initial load's last item. Since future loads are based on this initial lastAccessedIndex, it essentially acts as the initial "anchor" and so we want to match its computation with the computation of real paging's initial "anchor".

Test: ./gradlew paging:paging-testing:test
Bug: 235528239
Change-Id: I654e3a142c702b93a6df8f85bb73860b10e20d99