Introduce Impulse Strategy and Differential Values to VelocityTracker1D
This allows to reliably calculate differential velocities (i.e.
velocities for data points provided as differences between consequetive
values, instead of absolute/positional values). The logic is based on
Android's implementation of the Impulse velocity tracking strategy. Some
tests from Android has been ported to verify the logic's implementation
here.

In addition to Android, the initial attempt to introduce Impulse
velocity tracking to 2D/planar motions has also been referred to, for
instance, for the documentation of the Impulse calculation rationale
(see ref [1]).

This implementation of Impulse strategy fixes a small bug in the CL
provided in [1], where the duration to assume a pointer has stopped was
not respected (for instance - see `threePoints2` test case, where the
last data point comes 40+ms after the second-to-last. As such, all data
points except the last one should have been invalid, and since there's
only one valid data point, the velocity should have been zero). This
implementation fixes that bug. Some of the tests from that change have
been moved here.

Note that: differential velocity calculation has not been supported in
this change, as we have no need to do so right now - Impulse is
preferred for differential cases. If we decide to switch to Impulse for
2D/planar axes in the future, we would only have to adjust the
constructor params for the 1D-trackers initialized in the
VelocityTracker class.

[1] https://android-review.git.corp.google.com/c/platform/frameworks/support/+/2043280

Test: VelocityTracker1DTest, VelocityTrackerTest
Bug: 246395861

Change-Id: Ic657a0f080a696a339b0477ddd350c1ecadbdd13