Refactor MutableCombinedLoadStateCollection to expose StateFlow
This class used to expose its internal StateFlow<CombinedLoadStates?> as a Flow<CombinedLoadState>. We now consolidate the two flows into a single StateFlow<CombinedLoadStates?> such that the UI listeners can choose to use it as either a Flow or StateFlow.

This is part of the effort to enable Compose with initial LoadStates.

Test: ./gradlew paging:paging-common:test
Bug: 262441182
Change-Id: I36d8b0c70ca9f0fd90bc488cb28045cd25c6ea30