Graduate experimental LineBreak and Hyphens APIs to stable
This CL also includes an overall cleanup of `@OptIn(ExperimentalTextApi::class)` usage in `TextStyleTest` and `ParagraphStyleTest`.

Note: The `updateApi` build was failing because of a known bug in Metalava (b/244408981) which should have been fixed but it still affects the current change. This change was updated with `ignoreApiChange && updateApi`.

Test: :compose:ui:ui-text:cAT
Relnote: "LineBreak and Hyphens APIs in TextStyle are graduated to stable."
Change-Id: Ic1e1d5f43e0601251aa3cac549cd45b7cbb70aee