Fix method ordering for overloaded methods.
This CL fixes an ordering issue with overloaded methods. The previous
implementation tried to order methods purely by name, which then fails
for overloaded methods since they have the same name. The new
implementation takes the entire method descriptor into account, which
should avoid the ambiguity for overloaded methods.

Note that the majority of this CL is dedicated to
`KspJvmDescriptorUtils` and its corresponding
`KspJvmDescriptorUtilsTest`. In particular, I needed to add a new
interface to represent type variables, `XTypeVariableType`, in order to
handle type variables in a compiler agnostic way. There was also a bit
of trickiness to represent intersection types consistently across both
Javac and KSP (see the javadoc for `XTypeVariableType#upperBounds()`).

Test: KspJvmDescriptorTest and XTypeElementTest
Change-Id: I01f165ef6f2128884745f4da90d280518938ccca