AndroidX Webkit: reduce margins for medium interstitials
No change to logic. This reduces the margins around the WebViews in the
MediumInterstitialActivity because these WebViews are too small on
devices with a slimmer screen. If the WebViews are too narrow, then this
triggers "small" interstitials rather than "medium" interstitials, which
is not what this Activity is supposed to demonstrate.

This is a speculative fix for b/263499208.

Fixes: 263499208
Test: None - this is speculative
Change-Id: Ied5acffebcc128484c502c3c4c137c93759aad8f