Fix Robolectric Compose integration.
With the new Compose test runtime changes introduced in aosp/2260389 and
aosp/2243539, tests ran with Robolectric could get into a state where
waitForIdle would return before idle was actually achieved. If an effect
posted some work to the main thread, but compose itself was idle, then
that main thread work wouldn't get drained until the next call to
waitForIdle. Because we ran Espresso's onIdle before checking compose's
idle state, we wouldn't catch that extra main thread work.

This fix checks compose's idle state both before and after running
Espresso's onIdle, and continues to try to drain the main queue if
compose was not idle in either of those cases.

This was caught by a test for popups, so @aelias added a test for that,
and I also added a test for the lower-level queueing case.

Bug: 261739983
Test: RobolectricComposeTest.{testComposePostsToMainThread, testWaitForPopupWindow}
Test: ./gradlew :compose:ui:ui-test-junit4:testDebugUnitTest --tests androidx.compose.ui.test.junit4.RobolectricComposeTest
Change-Id: Ibff312a7454abf3fecf1f643bcf69eb8b8688f96