Ensure all ActiveCameras close camera eventually
Currently in the request loop of VirtualCameraManager, if the camera to
open isn't already opened, we would attempt to open the camera, but only
connect the VirtualCamera to AndroidCameraState (with token) in the next
iteration of the request loop.

This causes an issue because in every iteration of the request loop, we
always check whether the state of the VirtualCamera is
CameraStateUnopened. If for any reason, the camera state changes between
the current iteration (that makes the open attempt), and the next
iteration (that connects the VirtualCamera), we won't connect the
VirtualCamera that resides in the initial open request.

More generally speaking, the lifetime of cameras should be completely
tied to the creation of ActiveCamera. In this CL, we make sure the
WakeLock ActiveCameras does get triggered eventually by having
WakeLockToken start the countdown timeout immediately when the WakeLock
is created.

Bug: 264396089
Test: With the following:
  1. Make openCameraWithRetry() take longer
  2. Open CameraPipe test app
  3. Go to home screen (trigger onStop()) before camera open finishes
  4. Verify that the camera is still closed in the end.
Change-Id: I5b41b6a5dfee0a98019de8c9af2e804d20537850