[Camera-pipe] Call onCameraControlReady after configure camera
The VideoCapture requires updating the SessionConfig in the
onStateAttached callback; we may need to call it before configuring
the camera.
On the other hand, ImageCapture may require updating the CameraControl
in the onCameraControlReady callback, so we should ensure the CameraControl
is available when calling it.

In this CL, we do not call onCameraControlReady in the onStateAttached.
In CameraPipe-integration, the onCameraControlReady will be called after
the UseCaseCamera become available, onStateAttached will be called
before configuring the camera.

Bug: 264633438
Test: VideoCaptureDeviceTest, FlashTest, TapToFocusDeviceTest
Change-Id: Ic9772f045f541a6ec798d6237e881d1371c284a0