Features now store indices to cubics, not cubics directly
Previously, the cubics inside of Feature objects (Corners and
Edges) could get out of sync with the cubics in the shape described
by the feature (because cubics can be transformed and Features cannot).
This change alters how we store the Feature information. Instead of each
Feature having a list of Cubic objects, it now has a list of integer
indices, which refer to the cubics in the shape's cubics list.

This means that any use of a Feature must use both the cubics list
and the per-Feature list of indices to get the necessary data to
reproduce the cubics for a given Feature.

This CL also fixes a bug where features were not being copied
in the Polygon copy constructor.

Bug: 223586753
Test: Added test for features ad passed all tests
Change-Id: Iaafcb8a5a6dc509c9513a80a3c42ebf15e8c264e