Add superTypes to AnnotatedInterface
I chose to add this to AnnotatedInterface instead of Type as it is less
intrusive, and we currently do not have a usecase for general Type
supertype information. Let me know if you think it should be on Type
instead.

Since we are now recording supertype informaton, I have filtered out Any
as a default supertype so that we do not have to consider the Any case
in all future code gen. However, this makes us a bit inconsistent with
KSP.

Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,apipackager,core}:test --parallel
Bug: 265266769
Change-Id: I2c333d657efe9d4b03b2eb6954f1fff6730b554c