Fix ExistingActivityLifecycleTest failure with CameraPipeConfig
When binding UseCases with CameraPipeCOnfig, there is a race condition outside of camera-core so that the order of the UseCase in the collection matters.

Prior to aosp/2387898, CameraXActivity binds UseCases in the order of [Preview, ImageCapture, ImageAnalysis]. In aosp/2387898, the new implementation wraps the UseCases in a HashSet, which may change the order to [ImageCapture, ImageAnalysis, Preview]. This change causes the tests to fail.

In this CL, we replace HashSet with ArrayList to make sure CameraUseCaseAdapter maintain the order. However the isssue in the lower layer (likely camera-pipe-integration) will need seperate investigation. Please see aosp/2405518 on how to reproduce it.

Bug:266641900
Test: manual test and ./gradlew bOS
Change-Id: Ib575b49f62cf6b060c09f1524620f9b9e1477625