docs: Fixing summary fragments
The linter gave a warning about the typo "An transition" in both
files (should be "A transition"), but when I opened the file, I
saw several summary fragments that didn't end with a period;
that results in the whole description being pulled into the
summary table instead of just the summary fragment (see e.g.
https://screenshot.googleplex.com/3g6qe7M8tcUySUr.png).

Since I had the files open, I went ahead and fixed those fragments.
No changes to code, just to the doc comments.

Bug: 266857135
Test: [go/abtd docs build]
Change-Id: I18c1db4ccbc7d217e86a3a6ed6963e98e72ad313