Cast method call args in CVF fix if they don't match expected types
If X extends Y and a method expects type Y, when an X is used instead, there's an implicit cast to Y. If X and Y were introduced in different API levels, this implicit cast can cause verification issues because on certain API levels the verifier will see a cast from Object (instead of X) to Y, which is invalid. This change is to make that cast explicit for any CVF fix where the type of an argument used isn't identical to the type the method is expecting.

This adds a cast in all cases where the two types don't match, without checking if the implicit cast is between types from different API levels -- i.e. if X extends Y and X and Y were introduced in the same API level, an implicit cast from X to Y wouldn't actually cause a verification issue, but this adds the explicit cast anyway.

Test: Added a new test for this case, and a specific one for variable number of arguments, because that caused some errors.
Bug: 266845827
Change-Id: I022e38708ab19da0d26d2e2a23319970a632fc0a