Provide a way to obtain animation-independent bounds
Added a couple of methods in MotionLayoutScope to be able to obtain
final bounds of widgets in the scene.

Relnote: "Use `Modifier.onStartEndBoundsChanged(...)` to
obtain local bounds of Composables in a MotionLayout that are not
affected by animation. Useful for any ui interaction that requires
layout information that may also trigger animations such as DragAndDrop."

Bug: n/a
Test: MotionLayoutTest.testStartAndEndBoundsModifier
Change-Id: I6b5f9882acf81386dc62340b5f1ffb46ca91c750