Add CurvedBox component
CurvedBox places components on top of each other in the curved world. It is similar to what Box does but instead places them on an arc.

The thickness of the layout (the difference between the outer and inner radius) will be the same as the thickest child, and the angle taken will be the biggest angle of the children.

Bug: 269236803
Test: androidx.wear.compose.foundation.CurvedBoxTest.kt
Relnote: "Add CurvedBox component, which places components on top of each other in the curved world"
Change-Id: I29200406c7f7ea37d750ffcd8f478d3c8b2ea7b8