Use JVM reflection for modifier inspector defaults
This reverts a change from dbe4cb123ac8e15058a08154d67f229cde8273e5
where we used kotlin-reflect to provide an automatic population of
Modifier.Node default inspectable properties. This now uses an
expect/actual declaration, which should make the default implementation
more available in general, and indirectly fixes an issue in Play on G3
where the dependency on kotlin-reflect was falsely detected as a
transitive dependency and prevented shipping.

Fixes: b/269496185
Test: ./gradlew :compose:ui:ui:tDUT
Change-Id: I4cbd7d08f1ccabe5e50917a7280d3450ed6c89df