Revert "Added classname for LinearLayoutManager"
This reverts commit 27b68786bd9d0558231ebf92fa454cade57660db.

Reason for revert: Many breakages from tests (in apps and platform) from tests relying on the class name.  Also, Talkback already correctly recognizes one-column/one-row RecyclerViews as lists, so there's not a benefit making the change worth the issues.

Discussed and agreed on this solution in a sync.  For now, we are trying to just revert this for lists while leaving grids and staggered grids with the changed class names.

Bug: 261760077

Change-Id: Ie3ddd605a6bf5c80d08beb3d6a1b7accf65277ac