WIP BasicTextInputPlugin, BasicTextField2, TextFieldState
Adding a new TextInputPlugin to use with the new BasicTextField2. AndroidTextInputAdapter provides the necessary wiring to start a new input session and dispose it. Created InputConnections are not managed, but created input connections read and write to currently active input session through AndroidTextInputAdapter.

New text input sessions have a one-to-one relation with a `TextFieldState` to make sure that a single input session only modifies the given state. If TextFieldState argument reference changes for a TextField, text input session should be restarted.

Also added a very simple first implementation of BasicTextField2. This composable uses TextMeasurer for text layout instead of TextDelegate. It also means that we would like graduate TextMeasurer in the next release.

Test: N/A
Change-Id: Ia5b167241eccc724153446e977a50c702d8d8120