Try to fix flakiness of e2e tests
Many of these tests are failing regularly with Caller Not Authorized exceptions. This should be handled by overriding the allowlists in the test setup. From local testing, it seems that adding in a sleep after overriding the allowlist but before starting the test greatly reduces the chance of such a failure.

Test: Ran tests locally
Change-Id: Iff0873b0e8a19e177169dd592624c9a87fcc5509