Changed baseline profile configurations to use AGP attributes
This PR changes the configuration attributes used in producer and
consumer plugins, to utilize both BuildTypeAttr and ProductFlavorAttr
defined in AGP. It also splits the product flavor attributes per
dimension instead of using the final name, in order to allow also
partial matching if some dimensions are missing. Some other attributes
have been added to better define the content of the configurations and
improve the gradle configuration matching.

Test: Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Bug: 266214654
Change-Id: I74abb4446235fa2b0a08a35122d2b29e358dab78